+++ engine/src/render/voxels/palette.cpp
@@ -1,0 +1,10 @@
+#include <cubos/core/ecs/reflection.hpp>
+
+#include <cubos/engine/render/voxels/palette.hpp>
+
+CUBOS_REFLECT_IMPL(cubos::engine::RenderPalette)
+{
+    return core::ecs::TypeBuilder<RenderPalette>("cubos::engine::RenderPalette")
+        .withField("asset", &RenderPalette::asset)
+        .build();
+}